% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisBlast.R
\name{VisBlast}
\alias{VisBlast}
\title{Plot synteny comparison between two genomes using GFF and BLAST}
\usage{
VisBlast(
  query_gff,
  subject_gff,
  blast_tsv,
  output = "synteny.png",
  target_genes = c("geneA"),
  target_labels = c("geneA"),
  target_colors = c("#ec6800"),
  query_color = "#028760",
  subject_color = "#028760",
  query_label = "Query",
  subject_label = "Subject",
<<<<<<< HEAD
  swap_axes = FALSE,
=======
>>>>>>> 95e12c7 (Update VisBlast function and documentation (v2))
  plot_width = 10,
  plot_height = 6
)
}
\arguments{
\item{query_gff}{Path to query GFF file}

\item{subject_gff}{Path to subject GFF file}

\item{blast_tsv}{Path to BLAST tabular file (outfmt 6)}

\item{output}{Path to output PNG file, or NULL to skip saving}

\item{target_genes}{Vector of gene labels to highlight}

\item{target_labels}{Vector of labels to display}

\item{target_colors}{Vector of colors for highlighted BLAST bands}

\item{query_color}{Color for query gene arrows}

\item{subject_color}{Color for subject gene arrows}

\item{query_label}{Label to display at the left of the query line}

\item{subject_label}{Label to display at the left of the subject line}

<<<<<<< HEAD
\item{swap_axes}{Logical, if TRUE swap query and subject vertical positions}

=======
>>>>>>> 95e12c7 (Update VisBlast function and documentation (v2))
\item{plot_width}{Width of output PNG (default 10)}

\item{plot_height}{Height of output PNG (default 6)}
}
\value{
Saves a PNG plot (if output is specified) and returns ggplot object
}
\description{
Plot synteny comparison between two genomes using GFF and BLAST
}
